Refactor and fix type issues
