+++ client-next/src/components/ItineraryList/ItineraryHeaderContent.tsx
@@ -1,50 +1,25 @@
 import { TripPattern } from '../../gql/graphql.ts';
-import { useMemo } from 'react';
+import { TIME_WIDTH, useHeaderContentStyleCalculations } from './useHeaderContentStyleCalculations.ts';
 
 export function ItineraryHeaderContent({
-  containerWidth,
   tripPattern,
   itineraryIndex,
+  containerWidth,
   earliestStartTime,
   latestEndTime,
 }: {
-  containerWidth: number;
   tripPattern: TripPattern;
   itineraryIndex: number;
+  containerWidth: number;
   earliestStartTime: string | null;
   latestEndTime: string | null;
 }) {
-  const maxSpan = useMemo(
-    () => new Date(latestEndTime!).getTime() - new Date(earliestStartTime!).getTime(),
-    [earliestStartTime, latestEndTime],
+  const { widthPx, leftPx } = useHeaderContentStyleCalculations(
+    tripPattern,
+    containerWidth,
+    earliestStartTime,
+    latestEndTime,
   );
-
-  const startPct = useMemo(
-    () => (new Date(tripPattern.expectedStartTime).getTime() - new Date(earliestStartTime!).getTime()) / maxSpan,
-    [tripPattern.expectedStartTime, earliestStartTime],
-  );
-
-  const itinSpan = useMemo(
-    () => new Date(tripPattern.expectedEndTime).getTime() - new Date(tripPattern.expectedStartTime).getTime(),
-    [tripPattern.expectedStartTime, tripPattern.expectedEndTime],
-  );
-
-  const CONTAINER_WIDTH_PADDING = 90;
-  const paddedContainerWidth = containerWidth - CONTAINER_WIDTH_PADDING;
-
-  // var timeWidth = 40;
-  const timeWidth = 40;
-  // var startPx = 20 + timeWidth,
-  const startPx = 20 + timeWidth;
-  //   endPx = div.width() - timeWidth - (itin.groupSize ? 48 : 0);
-  const endPx = paddedContainerWidth - timeWidth;
-
-  // var pxSpan = endPx - startPx;
-  const pxSpan = endPx - startPx;
-  // var leftPx = startPx + startPct * pxSpan;
-  const leftPx = startPx + startPct * pxSpan;
-  // var widthPx = pxSpan * (itinSpan / maxSpan);
-  const widthPx = pxSpan * (itinSpan / maxSpan);
 
   return (
     <div style={{ position: 'relative' }}>
@@ -62,7 +37,7 @@ export function ItineraryHeaderContent({
       <div
         style={{
           position: 'absolute',
-          left: `${leftPx - timeWidth}px`,
+          left: `${leftPx - TIME_WIDTH}px`,
           background: 'black',
           color: 'white',
         }}
